requirements.txt
streamlit
google-generativeai>=0.8.3

python-dotenv
nltk

transformers

langchain
pandas

textblob


googletrans


gunicorn



python-decouple

